requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.35
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

pytest>=7.4.3
pytest-cov>=4.1.0
pytest-flask>=1.3.0

python-dateutil>=2.8.2
gunicorn>=21.2.0

